refactor: Ticket API 경계를 독립 패키지로 분리
Constraint: Ticket maker 조회/수정 흐름은 Booking aggregate command와 다른 실행 API 경계다.
Rejected: keeping ticket controller/service/codes under apis.booking | ticket use-case language already owns dedicated API/application contracts.
Confidence: high
Scope-risk: moderate
Directive: place future ticket API/application response and error codes under apis.ticket, not apis.booking or domain.booking.
Tested: ./gradlew :apis:compileKotlin :apis:compileJava :apis:compileTestKotlin :apis:compileTestJava --no-daemon
Tested: ./gradlew :apis:test --tests com.beat.apis.ticket.api.TicketControllerTest --tests com.beat.apis.boundary.DomainApplicationCodeBoundarySnapshotTest --no-daemon
Tested: ./gradlew :test --tests com.beat.SharedBoundaryContractTest --rerun-tasks --no-daemon
Tested: ./gradlew transitionBoundaryTest --no-daemon
Tested: ./gradlew check --no-daemon
Not-tested: production DB-backed ticket API smoke test
