+++ apis/src/main/java/com/beat/apis/ticket/application/dto/TicketRetrieveResponse.java
@@ -1,4 +1,4 @@
-package com.beat.apis.booking.application.dto;
+package com.beat.apis.ticket.application.dto;
 
 import java.util.List;
