+++ apis/src/main/kotlin/com/beat/apis/booking/application/exception/BookingApplicationErrorCode.kt
@@ -7,7 +7,6 @@ enum class BookingApplicationErrorCode(
     private val message: String,
 ) : BaseErrorCode {
     NO_BOOKING_FOUND(404, "입력하신 정보와 일치하는 예매 내역이 없습니다. 확인 후 다시 조회해주세요."),
-    NO_TICKETS_FOUND(404, "입력하신 정보와 일치하는 예매자 목록이 없습니다."),
     NO_PERFORMANCE_FOUND(404, "공연을 찾을 수 없습니다."),
     NO_SCHEDULE_FOUND(404, "회차를 찾을 수 없습니다.")
     ;
