+++ docs/migration/domain-application-errorcode-review-checklist.md
@@ -32,8 +32,8 @@ These are candidates that need special attention in the inventory because their 
 | --- | --- | --- | --- | --- |
 | `BookingErrorCode` | `INVALID_DATA_FORMAT` | thrown by `domain.booking.domain.Booking` | true booking invariant; should not be merged with request/query format codes of the same enum without review | `domain-owned` |
 | `BookingErrorCode` | `REQUIRED_DATA_MISSING`, `INVALID_REQUEST_FORMAT`, lookup `NO_*` codes | executable booking services | request/query and lookup-flow language is application-facing | `application-owned` unless a domain invariant throw site is found |
-| `BookingErrorCode` | `NO_TICKETS_FOUND`, `NO_PERFORMANCE_FOUND` | no main-code throw site found in this worktree | unused constants can hide stale API messages | `unused/defer`; do not move without consumer check |
-| `TicketErrorCode` | all constants | controller/service validation and ticket update flow | currently application/API policy, not a pure ticket domain model throw | `application-owned` unless follow-up introduces domain ticket model invariant use |
+| `BookingErrorCode` / `TicketApplicationErrorCode` | `NO_PERFORMANCE_FOUND` / `NO_TICKETS_FOUND` | `NO_PERFORMANCE_FOUND` has no main-code throw site; `NO_TICKETS_FOUND` is ticket response language | stale lookup messages can hide in the wrong context | keep ticket lookup language under `apis.ticket`; defer truly unused booking constants |
+| `TicketApplicationErrorCode` | all constants | controller/service validation and ticket update flow | moved to `apis.ticket.application.exception`; not a pure ticket domain model throw | `application-owned` unless follow-up introduces domain ticket model invariant use |
 | `MemberErrorCode` | `SOCIAL_TYPE_BAD_REQUEST` | `infra` Kakao adapter | external-provider/auth adapter failure does not belong to domain | `application/infra-auth-owned` follow-up needed |
 | `MemberErrorCode`, `UserErrorCode` | `*_NOT_FOUND` | executable services/facades | lookup failures are use-case/persistence orchestration, not domain invariants | `application-owned` |
 | `PerformanceErrorCode` | `NEGATIVE_TICKET_PRICE`, `INVALID_DATA_FORMAT` | `domain.performance.domain.Performance` | pure domain validation currently depends on these constants | `domain-owned` |
